tracing capability in assembly, main to separate file, all start routines to test app
